moved threading utils into core package
